% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pencil.R
\name{pencil}
\alias{pencil}
\title{Calculates Pencil}
\usage{
pencil(lead, tree, paint)
}
\arguments{
\item{lead}{mass of lead (grams)}

\item{tree}{number of tree shavings}

\item{paint}{amount of paint (ounces)}
}
\description{
Calculates Pencil
}
\examples{
pencil(lead = 4, tree = 8, paint = 6)
}
